--- hivespace.microservice.slnLaunch.user
@@ -1,11 +1,0 @@
-[
-  {
-    "Name": "New Profile",
-    "Projects": [
-      {
-        "Path": "src\\HiveSpace.ApiGateway\\HiveSpace.YarpApiGateway\\HiveSpace.YarpApiGateway.csproj",
-        "Action": "Start"
-      }
-    ]
-  }
-]
