+++ src/HiveSpace.ApiGateway/HiveSpace.YarpApiGateway/appsettings.json
@@ -8,7 +8,7 @@
   "AllowedHosts": "*",
   "ReverseProxy": {
     "Routes": {
-      "user-route": {
+      "identity-route": {
         "ClusterId": "user-cluster",
         "Match": {
           "Path": "/identity/{**catch-all}"
@@ -18,6 +18,18 @@
             "PathRemovePrefix": "/identity"
           }
         ]
+      },
+      "user-route": {
+        "ClusterId": "user-cluster",
+        "Match": {
+          "Path": "/api/v{version}/users/{**catch-all}"
+        }
+      },
+      "admin-route": {
+        "ClusterId": "user-cluster",
+        "Match": {
+          "Path": "/api/v{version}/admins/{**catch-all}"
+        }
       }
     },
     "Clusters": {
