+++ src/HiveSpace.UserService/HiveSpace.UserService.Application/Models/Responses/Admin/CreateAdminResponseDto.cs
@@ -3,7 +3,8 @@ namespace HiveSpace.UserService.Application.Models.Responses.Admin;
 public record CreateAdminResponseDto(
     Guid Id,
     string Email,
-    string UserName,
     string FullName,
-    string Role
+    bool IsSystemAdmin,
+    DateTimeOffset CreatedAt,
+    bool IsActive = true
 );
