+++ src/HiveSpace.UserService/HiveSpace.UserService.Domain/HiveSpace.UserService.Domain.csproj
@@ -8,7 +8,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\libs\HiveSpace.Domain.Shared\HiveSpace.Domain.Shared.csproj" />
-    <ProjectReference Include="..\..\..\libs\HiveSpace.Core\HiveSpace.Core.csproj" />
   </ItemGroup>
 
   <ItemGroup>
